feat: server works, move models to models pkg for avoid same generated name
