+++ rpc/triple/pb2/api/helloworld.proto
@@ -1,0 +1,30 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.dubbogo.samples.rpc.triple.pb2.api;
+
+import "github.com/apache/dubbo-go-samples/rpc/triple/pb2/models/generated.proto";
+
+option go_package = "./;api";
+
+// The greeting service definition.
+service Greeter {
+  // Sends a greeting
+  rpc SayHello (github.com.apache.dubbo_go_samples.rpc.triple.pb2.models.HelloRequest) returns (github.com.apache.dubbo_go_samples.rpc.triple.pb2.models.User) {}
+  // Sends a greeting via stream
+  rpc SayHelloStream (stream github.com.apache.dubbo_go_samples.rpc.triple.pb2.models.HelloRequest) returns (stream github.com.apache.dubbo_go_samples.rpc.triple.pb2.models.User) {}
+}
