+++ rpc/triple/pb2/go-server/cmd/server.go
@@ -1,20 +1,3 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 package main
 
 import (
@@ -31,18 +14,15 @@ import (
 
 	"github.com/dubbogo/gost/log/logger"
 
+	"github.com/apache/dubbo-go-samples/rpc/triple/pb2/go-server/pkg"
 	_ "github.com/dubbogo/triple/pkg/triple"
-)
-
-import (
-	"github.com/apache/dubbo-go-samples/rpc/triple/pb/dubbogo-grpc/go-server/pkg"
 )
 
 var (
 	survivalTimeout = int(3 * time.Second)
 )
 
-// export DUBBO_GO_CONFIG_PATH=$PATH_TO_SAMPLES/rpc/triple/pb2/go-server/dubbogo-server/conf/dubbogo.yml
+// export DUBBO_GO_CONFIG_PATH=$PATH_TO_SAMPLES/rpc/triple/pb2/go-server/conf/dubbogo.yml
 func main() {
 	config.SetProviderService(&pkg.GreeterProvider{})
 	if err := config.Load(); err != nil {
