+++ rpc/triple/pb2/go-server/conf/dubbogo.yml
@@ -1,0 +1,10 @@
+dubbo:
+  protocols:
+    triple:
+      name: tri
+      port: 20000
+  provider:
+    services:
+      GreeterProvider:
+        # interface read from pb file
+        interface: ""
